feat(FN-3633): rename file to match helper function name
