+++ portal/server/helpers/filterReportJsonToRelevantKeys.js
